+++ openmdao/drivers/scipy_optimizer.py
@@ -333,7 +333,7 @@ class ScipyOptimizeDriver(Driver):
 
         # Constraints
         constraints = []
-        i = 1  # start at 1 since row 0 is the objective.  Constraints start at row 1.
+        nl_i = 1  # start at 1 since row 0 is the objective.  Constraints start at row 1.
         lin_i = 0  # counter for linear constraint jacobian
         lincons = []  # list of linear constraints
         self._obj_and_nlcons = list(self._objs)
@@ -367,8 +367,8 @@ class ScipyOptimizeDriver(Driver):
                     lin_i += size
                 else:
                     self._obj_and_nlcons.append(name)
-                    self._con_idx[name] = i
-                    i += size
+                    self._con_idx[name] = nl_i
+                    nl_i += size
 
                 # In scipy constraint optimizers take constraints in two separate formats