+++ openmdao/utils/tests/test_reports_system.py
@@ -5,6 +5,8 @@ import sys
 import os
 from io import StringIO
 
+import numpy as np
+
 import openmdao.api as om
 from openmdao.test_suite.components.paraboloid import Paraboloid
 from openmdao.test_suite.groups.parallel_groups import Diamond
@@ -12,8 +14,8 @@ from openmdao.core.problem import _default_prob_name
 import openmdao.core.problem as probmod
 from openmdao.core.constants import _UNDEFINED
 from openmdao.utils.general_utils import set_pyoptsparse_opt
-from openmdao.utils.reports_system import _reports_dir, register_report, \
-    list_reports, clear_reports, _reset_reports_dir, activate_report, _reports_registry
+from openmdao.utils.reports_system import register_report, \
+    list_reports, clear_reports, activate_report, _reports_registry
 from openmdao.utils.testing_utils import use_tempdirs
 from openmdao.utils.assert_utils import assert_no_warning
 from openmdao.utils.mpi import MPI
@@ -43,13 +45,11 @@ class TestReportsSystem(unittest.TestCase):
         # set things to a known initial state for all the test runs
         probmod._clear_problem_names()  # need to reset these to simulate separate runs
         os.environ.pop('OPENMDAO_REPORTS', None)
-        os.environ.pop('OPENMDAO_REPORTS_DIR', None)
         # We need to remove the TESTFLO_RUNNING environment variable for these tests to run.
         # The reports code checks to see if TESTFLO_RUNNING is set and will not do anything if set
         # But we need to remember whether it was set so we can restore it
         self.testflo_running = os.environ.pop('TESTFLO_RUNNING', None)
         clear_reports()
-        om.set_reports_dir(_reports_dir)
 
         self.count = 0
 
@@ -59,9 +59,6 @@ class TestReportsSystem(unittest.TestCase):
             os.environ['TESTFLO_RUNNING'] = self.testflo_running
 
     def setup_and_run_simple_problem(self, driver=None, reports=_UNDEFINED, reports_dir=_UNDEFINED, linear=False):
-        if reports_dir is not _UNDEFINED:
-            om.set_reports_dir(reports_dir)
-
         prob = om.Problem(reports=reports)
         model = prob.model
 
@@ -89,10 +86,48 @@ class TestReportsSystem(unittest.TestCase):
 
         return prob
 
-    def setup_problem_w_errors(self, prob_name, driver=None, reports=_UNDEFINED, reports_dir=_UNDEFINED):
-        if reports_dir is not _UNDEFINED:
-            om.set_reports_dir(reports_dir)
+    def setup_and_run_w_linear_only_dvs(self, driver, reports=_UNDEFINED, reports_dir=_UNDEFINED, shape=3):
+        prob = om.Problem(reports=reports)
+        prob.driver = driver
+        model = prob.model
 
+        ivc = model.add_subsystem('ivc', om.IndepVarComp())
+        ivc.add_output('x', np.ones(shape))
+        ivc.add_output('y', np.ones(shape))
+        ivc.add_output('z', np.ones(shape))
+
+        model.add_subsystem('comp', om.ExecComp('f_xy=x*2.-y*3.', shape=shape))
+        model.add_subsystem('obj', om.ExecComp('obj = sum(x**2)', obj=1., x=np.ones(shape)))
+        model.add_subsystem('con', om.ExecComp('y=x', shape=shape))
+        model.add_subsystem('con2', om.ExecComp('y=sin(x)', shape=shape))
+        model.add_subsystem('con3', om.ExecComp('y=.2*x', shape=shape), promotes_inputs=['x'])
+        model.add_subsystem('con4', om.ExecComp('y=cos(x)', shape=shape), promotes_inputs=['x'])
+
+        model.connect('ivc.x', 'comp.x')
+        model.connect('ivc.y', 'comp.y')
+        model.connect('comp.f_xy', 'obj.x')
+        model.connect('ivc.z', 'con.x')
+        model.connect('ivc.x', 'con2.x')
+
+        model.add_design_var('ivc.x', lower=0.0, upper=1.0)
+        model.add_design_var('ivc.y', lower=0.0, upper=1.0)
+        model.add_design_var('ivc.z', lower=0.0, upper=1.0)
+        model.add_design_var('x', lower=0.0, upper=1.0)
+
+        model.add_objective('obj.obj')
+
+        model.add_constraint('con.y', lower=0.0, linear=True)
+        model.add_constraint('con3.y', lower=0.0, linear=True)
+        model.add_constraint('con4.y', lower=0.0)
+        model.add_constraint('con2.y', lower=0.0)
+
+        prob.setup(check=False)
+        prob.run_driver()
+        prob.cleanup()
+
+        return prob
+
+    def setup_problem_w_errors(self, prob_name, driver=None, reports=_UNDEFINED, reports_dir=_UNDEFINED):
         prob = om.Problem(reports=reports, name=prob_name)
         model = prob.model
 
@@ -148,14 +183,14 @@ class TestReportsSystem(unittest.TestCase):
         prob.setup()
         prob.run_model()  # need to do run_model in this test so sub problem is created
 
-        return prob._name, G.nonlinear_solver._problem._name
+        return prob, G.nonlinear_solver._problem
 
     @hooks_active
     def test_report_generation_basic(self):
         prob = self.setup_and_run_simple_problem()
 
         # get the path to the problem subdirectory
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
@@ -165,15 +200,51 @@ class TestReportsSystem(unittest.TestCase):
         self.assertTrue(path.is_file(), f'The optimizer report file, {str(path)}, was not found')
 
     @hooks_active
+    def test_report_generation_linear_only_dv_scaling_report_pyoptsparse(self):
+        if not OPTIMIZER:
+            raise unittest.SkipTest("This test requires pyOptSparseDriver.")
+
+        driver = om.pyOptSparseDriver(optimizer='IPOPT')
+        driver.declare_coloring()
+
+        driver.opt_settings['max_iter'] = 1000
+        driver.opt_settings['print_level'] = 5
+        driver.opt_settings['mu_strategy'] = 'monotone'
+        driver.opt_settings['alpha_for_y'] = 'safer-min-dual-infeas'
+        driver.opt_settings['tol'] = 1.0E-4
+        driver.opt_settings['constr_viol_tol'] = 1.0E-4
+
+        prob = self.setup_and_run_w_linear_only_dvs(driver=driver, reports=['scaling'], shape=(9,7))
+
+        # get the path to the problem subdirectory
+        problem_reports_dir = prob.get_reports_dir()
+
+        path = pathlib.Path(problem_reports_dir).joinpath(self.scaling_filename)
+        self.assertTrue(path.is_file(), f'The scaling report file, {str(path)} was not found')
+
+    @hooks_active
+    def test_report_generation_linear_only_dv_scaling_report_scipyopt(self):
+        driver = om.ScipyOptimizeDriver(optimizer='SLSQP')
+        driver.declare_coloring()
+
+        prob = self.setup_and_run_w_linear_only_dvs(driver=driver, reports=['scaling'], shape=(9,7))
+
+        # get the path to the problem subdirectory
+        problem_reports_dir = prob.get_reports_dir()
+
+        path = pathlib.Path(problem_reports_dir).joinpath(self.scaling_filename)
+        self.assertTrue(path.is_file(), f'The scaling report file, {str(path)} was not found')
+
+    @hooks_active
     def test_report_generation_on_error(self):
         prob_name = 'error_problem'
         try:
             prob = self.setup_problem_w_errors(prob_name)
         except Exception as err:
             # get the path to the problem subdirectory
-            problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob_name)
+            problem_reports_dir = pathlib.Path(f'{prob_name}_out/reports')
 
-            path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
+            path = problem_reports_dir / self.n2_filename
             self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
 
             self.assertEqual(str(err),
@@ -190,7 +261,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem(driver=pyOptSparseDriver(optimizer='SLSQP'))
 
         # get the path to the problem subdirectory
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
@@ -212,7 +283,7 @@ class TestReportsSystem(unittest.TestCase):
         # Test a driver that does not generate scaling report
         prob = self.setup_and_run_simple_problem(driver=om.DOEDriver(doe_list))
 
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
@@ -251,7 +322,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem()
 
         # See if the report files exist and if they have the right names
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertFalse(path.is_file(),
@@ -272,7 +343,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem()
 
         # See if the report files exist and if they have the right names
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
@@ -292,7 +363,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem(reports=['optimizer', 'scaling'])
 
         # See if the report files exist and if they have the right names
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertFalse(path.is_file(),
@@ -311,7 +382,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem(reports=False)
 
         # See if the report files exist and if they have the right names
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertFalse(path.is_file(),
@@ -322,26 +393,6 @@ class TestReportsSystem(unittest.TestCase):
         path = pathlib.Path(problem_reports_dir).joinpath(self.optimizer_filename)
         self.assertFalse(path.is_file(),
                          f'The optimizer report file, {str(path)}, was found but should not exist.')
-
-    @hooks_active
-    def test_report_generation_set_reports_dir_using_env_var(self):
-        # test use of setting a custom reports directory other than the default of "."
-        custom_dir = 'custom_reports_dir'
-        os.environ['OPENMDAO_REPORTS_DIR'] = custom_dir
-        _reset_reports_dir()  # this will use current value of OPENMDAO_REPORTS_DIR
-
-        prob = self.setup_and_run_simple_problem()
-
-        # See if the report files exist and if they have the right names
-        reports_dir = custom_dir
-        problem_reports_dir = pathlib.Path(reports_dir).joinpath(prob._name)
-
-        path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
-        self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
-        path = pathlib.Path(problem_reports_dir).joinpath(self.scaling_filename)
-        self.assertTrue(path.is_file(), f'The scaling report file, {str(path)}, was not found')
-        path = pathlib.Path(problem_reports_dir).joinpath(self.optimizer_filename)
-        self.assertTrue(path.is_file(), f'The optimizer report file, {str(path)}, was not found')
 
     @hooks_active
     def test_report_generation_user_defined_report(self):
@@ -355,11 +406,11 @@ class TestReportsSystem(unittest.TestCase):
 
         register_report("User report", user_defined_report,
                         "user report description",
-                        'Problem', 'setup', 'pre', report_filename=user_report_filename)
+                        'Problem', 'setup', 'post', report_filename=user_report_filename)
 
         prob = self.setup_and_run_simple_problem()
 
-        path = pathlib.Path(_reports_dir).joinpath(prob._name, user_report_filename)
+        path = prob.get_reports_dir() / user_report_filename
 
         self.assertTrue(path.is_file(), f'The user report file, {str(path)} was not found')
 
@@ -367,6 +418,24 @@ class TestReportsSystem(unittest.TestCase):
         self.assertTrue('User report' in _reports_registry, "'User report' not found in registry.")
         om.unregister_report('User report')
         self.assertFalse('User report' in _reports_registry, "'User report' found in registry.")
+
+    @hooks_active
+    def test_register_report_pre_setup(self):
+        user_report_filename = 'user_report.txt'
+        os.environ['OPENMDAO_REPORTS'] = 'User report'
+
+        def user_defined_report(prob, report_filename):
+            path = pathlib.Path(prob.get_reports_dir()).joinpath(report_filename)
+            with open(path, "w") as f:
+                f.write(f"Do some reporting on the Problem, {prob._name}\n")
+
+        with self.assertRaises(ValueError) as e:
+            register_report("User report", user_defined_report,
+                            "user report description",
+                            'Problem', 'setup', 'pre', report_filename=user_report_filename)
+
+        expected = 'Reports cannot be registered to execute pre-setup.'
+        self.assertEqual(str(e.exception), expected)
 
     @hooks_active
     def test_report_generation_various_locations(self):
@@ -386,6 +455,8 @@ class TestReportsSystem(unittest.TestCase):
 
         for method in ['setup', 'final_setup', 'run_driver']:
             for pre_or_post in ['pre', 'post']:
+                if (method, pre_or_post) == ('setup', 'pre'):
+                    continue
                 repname = f"User defined report {method} {pre_or_post}"
                 register_report(repname, user_defined_report,
                                 "user defined report", 'Problem', method, pre_or_post,
@@ -395,8 +466,10 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem()
 
         self.count = 0
-        for _ in ['setup', 'final_setup', 'run_driver']:
-            for _ in ['pre', 'post']:
+        for method in ['setup', 'final_setup', 'run_driver']:
+            for pre_or_post in ['pre', 'post']:
+                if (method, pre_or_post) == ('setup', 'pre'):
+                    continue
                 user_report_filename = f"user_defined_{self.count}.txt"
                 path = pathlib.Path(prob.get_reports_dir()).joinpath(user_report_filename)
                 self.assertTrue(path.is_file(),
@@ -405,11 +478,11 @@ class TestReportsSystem(unittest.TestCase):
 
     @hooks_active
     def test_report_generation_multiple_problems(self):
-        probname, subprobname = self.setup_and_run_model_with_subproblem()
+        prob, subprob = self.setup_and_run_model_with_subproblem()
 
         # The multiple problem code only runs model so no scaling or optimizer reports to look for
-        for problem_name in [probname, subprobname]:
-            problem_reports_dir = pathlib.Path(_reports_dir).joinpath(f'{problem_name}')
+        for p in [prob, subprob]:
+            problem_reports_dir = p.get_reports_dir()
             path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
             self.assertTrue(path.is_file(), f'N2 report file, {str(path)} was not found')
 
@@ -425,19 +498,19 @@ class TestReportsSystem(unittest.TestCase):
                         report_filename=self.n2_filename,
                         inst_id=_default_prob_name() + '2')
 
-        probname, subprobname = self.setup_and_run_model_with_subproblem()
+        prob, subprob = self.setup_and_run_model_with_subproblem()
 
-        self.assertEqual(_default_prob_name(), probname)
+        self.assertEqual(_default_prob_name(), prob._name)
 
         # The multiple problem code only runs model so no scaling reports to look for
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(f'{subprobname}')
-        path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
+        problem_reports_dir = subprob.get_reports_dir()
+        path = problem_reports_dir / self.n2_filename
         # for the subproblem named problem2, there should be a report but not for problem1 since
         #    we specifically asked for just the instance of problem2
         self.assertTrue(path.is_file(), f'The n2 report file, {str(path)} was not found')
 
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(f'{probname}')
-        path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
+        problem_reports_dir = prob.get_reports_dir()
+        path = problem_reports_dir / self.n2_filename
         self.assertFalse(path.is_file(),
                          f'The N2 report file, {str(path)} was found but should not exist.')
 
@@ -450,7 +523,7 @@ class TestReportsSystem(unittest.TestCase):
 
         prob = self.setup_and_run_simple_problem()
 
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertFalse(path.is_file(),
@@ -467,7 +540,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem(reports=False)
 
         # get the path to the problem subdirectory
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertFalse(path.is_file(),
                          f'The N2 report file, {str(path)} was found but should not exist.')
@@ -483,7 +556,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem(reports=None)
 
         # get the path to the problem subdirectory
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertFalse(path.is_file(),
@@ -500,7 +573,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem(reports='n2')
 
         # get the path to the problem subdirectory
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
@@ -516,7 +589,7 @@ class TestReportsSystem(unittest.TestCase):
         prob = self.setup_and_run_simple_problem(reports=['n2','scaling'])
 
         # get the path to the problem subdirectory
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
         self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')
@@ -528,18 +601,16 @@ class TestReportsSystem(unittest.TestCase):
 
     @hooks_active
     def test_report_generation_problem_reports_argument_multiple_problems(self):
-        _, _ = self.setup_and_run_model_with_subproblem(prob2_reports=None)
+        prob, subprob = self.setup_and_run_model_with_subproblem(prob2_reports=None)
 
         # Only problem1 reports should have been generated
 
         # The multiple problem code only runs model so no scaling reports to look for
-        problem_name = _default_prob_name()
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(f'{problem_name}')
-        path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
+        problem_reports_dir = prob.get_reports_dir()
+        path = problem_reports_dir / self.n2_filename
         self.assertTrue(path.is_file(), f'The problem1 N2 report file, {str(path)} was not found')
 
-        problem_name = problem_name + '2'
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(f'{problem_name}')
+        problem_reports_dir = subprob.get_reports_dir()
         self.assertFalse(problem_reports_dir.is_dir(),
                          f'The problem2 report dir was found but should not exist.')
         path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
@@ -573,7 +644,7 @@ class TestReportsSystem(unittest.TestCase):
         self.assertEqual(prob.driver.result.deriv_evals, 3)
 
         # See if the report files exist and if they have the right names
-        problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+        problem_reports_dir = prob.get_reports_dir()
 
         path = pathlib.Path(problem_reports_dir).joinpath(self.scaling_filename)
         self.assertTrue(path.is_file(), f'The scaling report file, {str(path)} was not found')
@@ -593,14 +664,12 @@ class TestReportsSystemMPI(unittest.TestCase):
         probmod._clear_problem_names()  # need to reset these to simulate separate runs
 
         os.environ.pop('OPENMDAO_REPORTS', None)
-        os.environ.pop('OPENMDAO_REPORTS_DIR', None)
         # We need to remove the TESTFLO_RUNNING environment variable for these tests to run.
         # The reports code checks to see if TESTFLO_RUNNING is set and will not do anything
         # if it is set.
         # But we need to remember whether it was set so we can restore it
         self.testflo_running = os.environ.pop('TESTFLO_RUNNING', None)
         clear_reports()
-        om.set_reports_dir(_reports_dir)
 
         self.count = 0  # used to keep a count of reports generated
 
@@ -625,7 +694,7 @@ class TestReportsSystemMPI(unittest.TestCase):
         prob.run_driver()
 
         if prob.comm.rank == 0:
-            problem_reports_dir = pathlib.Path(_reports_dir).joinpath(prob._name)
+            problem_reports_dir = prob.get_reports_dir()
 
             path = pathlib.Path(problem_reports_dir).joinpath(self.n2_filename)
             self.assertTrue(path.is_file(), f'The N2 report file, {str(path)} was not found')